Add FP16 support

+++ fairseq/models/fairseq_decoder.py
@@ -21,7 +21,7 @@ class FairseqDecoder(nn.Module):
 
     def get_normalized_probs(self, net_output, log_probs):
         """Get normalized probabilities (or log probs) from a net's output."""
-        logits = net_output[0]
+        logits = net_output[0].float()
         if log_probs:
             return F.log_softmax(logits, dim=-1)
         else:
